requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
python-dotenv
streamlit
requests
spacy
scikit-learn
PyPDF2
python-docx
python-multipart
